requirements.txt
pywin32
pandas
mysql-connector-python